# openapi.yaml
openapi: 3.0.3
info:
  title: TrustChecker Enterprise API
  version: 1.0.0
  description: |
    Supply Chain Trust Intelligence Platform API.
    TrustChecker provides trust scoring, supply chain operations,
    compliance automation, and carbon tracking for enterprise customers.
  contact:
    name: TrustChecker Support
    email: support@trustchecker.io
  license:
    name: Proprietary
servers:
  - url: https://tonytran.work/api/v1
    description: Production
  - url: http://localhost:4000/api
    description: Development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SSOConfig:
      type: object
      properties:
        provider: { type: string, enum: [saml, oauth2, oidc] }
        issuer_url: { type: string }
        sso_url: { type: string }
        enabled: { type: boolean }
security:
  - bearerAuth: []
paths:
...
  /sso/config:
    get:
      tags: [SSO]
      summary: Get SSO configuration
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  config: { $ref: "#/components/schemas/SSOConfig" }
    put:
      tags: [SSO]
      summary: Configure SSO for organization
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SSOConfig" }
      responses:
        200: { description: SSO configured }
  /sso/login/{orgSlug}:
    get:
      tags: [SSO]
      summary: Initiate SSO login (redirects to IdP)
      security: []
      parameters:
        - name: orgSlug
          in: path
          required: true
          schema: { type: string }
      responses:
        302: { description: Redirect to IdP }
...