# server/routes/sso.js
/**
 * SSO/SAML Enterprise Authentication v1.0
 * GET  /api/sso/config           — get SSO config for org
 * PUT  /api/sso/config           — update SSO config
 * GET  /api/sso/login/:org       — initiate SSO login
 * POST /api/sso/callback         — SSO callback (SAML ACS / OAuth2 callback)
 * GET  /api/sso/metadata/:org    — SAML metadata
 */
const express = require("express");
const router = express.Router();
const crypto = require("crypto");
const jwt = require("jsonwebtoken");
const db = require("../db");
const { authMiddleware } = require("../auth");

// Get SSO config (authenticated)
router.get("/config", authMiddleware, async function(req, res) {
    try {
        var result = await db.all(
            "SELECT id, org_id, provider, issuer_url, sso_url, metadata_url, attribute_mapping, enabled, enforce, created_at FROM sso_configs WHERE org_id = $1",
            [req.user.org_id]
        );
        res.json({ config: result[0] || null });
    } catch (err) {
        res.status(500).json({ error: "Failed to load SSO config" });
    }
});

// Update SSO config (admin only)
router.put("/config", authMiddleware, async function(req, res) {
    try {
        var b = req.body;
        if (!b.provider || !["saml", "oauth2", "oidc"].includes(b.provider)) {
            return res.status(400).json({ error: "provider must be saml, oauth2, or oidc" });
        }
        var result = await db.all(
            "INSERT INTO sso_configs (org_id, provider, issuer_url, sso_url, certificate, client_id, client_secret, metadata_url, attribute_mapping, enabled, enforce) VALUES ($1,$2,$3,$4,$5,$6,$7,$8,$9,$10,$11) ON CONFLICT (org_id) DO UPDATE SET provider=$2, issuer_url=$3, sso_url=$4, certificate=$5, client_id=$6, client_secret=$7, metadata_url=$8, attribute_mapping=$9, enabled=$10, enforce=$11, updated_at=NOW() RETURNING id",
            [req.user.org_id, b.provider, b.issuer_url, b.sso_url, b.certificate, b.client_id, b.client_secret, b.metadata_url, JSON.stringify(b.attribute_mapping || {}), b.enabled || false, b.enforce || false]
        );
        res.json({ id: result[0].id, status: "saved" });
    } catch (err) {
        res.status(500).json({ error: "Failed to save SSO config" });
    }
});

// Initiate SSO login (unauthenticated — redirects to IdP)
router.get("/login/:orgSlug", async function(req, res) {
    try {
        var result = await db.all(
            "SELECT sc.* FROM sso_configs sc JOIN organizations o ON o.id = sc.org_id WHERE o.slug = $1 AND sc.enabled = true",
            [req.params.orgSlug]
        );
        if (!result[0]) return res.status(404).json({ error: "SSO not configured for this organization" });
        var config = result[0];
        
        var state = crypto.randomBytes(32).toString("hex");
        // Store state in session/cookie for CSRF protection
        
        if (config.provider === "saml") {
            // SAML AuthnRequest redirect
            var samlRequest = Buffer.from("<samlp:AuthnRequest xmlns:samlp=\"urn:oasis:names:tc:SAML:2.0:protocol\" ID=\"_" + crypto.randomUUID() + "\" Version=\"2.0\" IssueInstant=\"" + new Date().toISOString() + "\" Destination=\"" + config.sso_url + "\" AssertionConsumerServiceURL=\"" + (process.env.BASE_URL || "http://localhost:4000") + "/api/sso/callback\"><saml:Issuer xmlns:saml=\"urn:oasis:names:tc:SAML:2.0:assertion\">" + (config.issuer_url || "trustchecker") + "</saml:Issuer></samlp:AuthnRequest>").toString("base64");
            return res.redirect(config.sso_url + "?SAMLRequest=" + encodeURIComponent(samlRequest) + "&RelayState=" + state);
        }
        
        if (config.provider === "oauth2" || config.provider === "oidc") {
            var authUrl = config.sso_url + "?client_id=" + config.client_id + "&response_type=code&redirect_uri=" + encodeURIComponent((process.env.BASE_URL || "http://localhost:4000") + "/api/sso/callback") + "&scope=openid%20email%20profile&state=" + state;
            return res.redirect(authUrl);
        }
        
        res.status(400).json({ error: "Unsupported provider" });
    } catch (err) {
        res.status(500).json({ error: "SSO initiation failed" });
    }
});

// SSO Callback (SAML ACS or OAuth2 callback)
router.post("/callback", async function(req, res) {
    try {
        // In production, validate SAML response signature or exchange OAuth2 code
        // For now, this is the framework — full SAML/OAuth2 validation requires passport-saml or similar
        var body = req.body;
        res.json({ 
            status: "callback_received",
            message: "Full SAML/OAuth2 validation requires passport-saml or openid-client package",
            integration_ready: true
        });
    } catch (err) {
        res.status(500).json({ error: "SSO callback failed" });
    }
});

// SAML metadata endpoint
router.get("/metadata/:orgSlug", async function(req, res) {
    var baseUrl = process.env.BASE_URL || "http://localhost:4000";
    var metadata = "<?xml version=\"1.0\"?><md:EntityDescriptor xmlns:md=\"urn:oasis:names:tc:SAML:2.0:metadata\" entityID=\"" + baseUrl + "/api/sso/metadata/" + req.params.orgSlug + "\"><md:SPSSODescriptor protocolSupportEnumeration=\"urn:oasis:names:tc:SAML:2.0:protocol\"><md:AssertionConsumerService Binding=\"urn:oasis:names:tc:SAML:2.0:bindings:HTTP-POST\" Location=\"" + baseUrl + "/api/sso/callback\" index=\"0\" isDefault=\"true\"/></md:SPSSODescriptor></md:EntityDescriptor>";
    res.type("application/xml").send(metadata);
});

module.exports = router;